fee now moved to EVM directory
